+++ src/app/services/validators.service.ts
@@ -28,17 +28,24 @@ export class ValidatorsService {
             errorMessage = 'Email Erróneo';
             break;
           case 'minlength':
-            errorMessage = `Mínimo ${errors['minlength']['requiredLength']} catacteres`;
+            errorMessage = `Mínimo ${errors['minlength']['requiredLength']} carácteres`;
             break;
           case 'isValidDni':
             errorMessage = `DNI inválido`;
             break;
           case 'notSamePassword':
-            errorMessage = `Las contraseñas no coinciden`
+            errorMessage = `Las contraseñas no coinciden`;
+            break;
+          case 'notChecked':
+            errorMessage = `Casilla obligatoria`;
+            break;
+          case 'invalidPassword':
+            errorMessage = `Mínimo 1 mayúscula, 1 minúscula, 1 número y sin caracteres especiales`;
             break;
           case 'pattern':
-            errorMessage = `Format incorrecto`
+            errorMessage = `Format0 incorrecto`;
             break;
+          
         }
       }
       return errorMessage;
@@ -61,7 +68,7 @@ export class ValidatorsService {
         return of(null);
       }
   
-      return of({ isValidDni: false });;
+      return of({ isValidDni: false });
     }
   
     public isSamePassword(password: string, confirmPassword: string) {
@@ -85,8 +92,15 @@ export class ValidatorsService {
       return of({notChecked: true});
     }
 
+    isValidPassword(control: AbstractControl): Observable<ValidationErrors | null>{
+      let password: string = control.value;
+      let passwordRegex = /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)[a-zA-Z0-9]+$/;
+  
+      if (passwordRegex.test(password)) {
+        return of(null);
+      }
+
+      return of({ invalidPassword: true });
+    }
+
 }
-
-
-
-