+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). 
 
+### [ita-challenges-frontend-2.0.52-RELEASE] - 2024-02-29
+* Fixing register and login basic functionalities 
+
 ### [ita-challenges-frontend-2.0.51-RELEASE] - 2024-02-29
 * Fixing nginx.conf to find assets folder